+++ src/EventStore.Core/Services/PersistentSubscription/IPersistentSubscriptionMessageParker.cs
@@ -7,8 +7,8 @@ namespace EventStore.Core.Services.PersistentSubscription
     public interface IPersistentSubscriptionMessageParker
     {
         void BeginParkMessage(ResolvedEvent ev, string reason, Action<ResolvedEvent, OperationResult> completed);
-        void BeginReadEndSequence(Action<int?> completed);
-        void BeginMarkParkedMessagesReprocessed(int sequence);
+        void BeginReadEndSequence(Action<long?> completed);
+        void BeginMarkParkedMessagesReprocessed(long sequence);
         void BeginDelete(Action<IPersistentSubscriptionMessageParker> completed);
     }
 }
